Renamed keyGen.to into keyGeneration.ts
